+++ benchmark/servers/index.mjs
@@ -1,5 +1,5 @@
 import './uWebSockets.mjs';
 import './ws8.mjs';
 import './ws7.mjs';
-import './fastify-websocket_ws8.mjs';
+import './fastify-websocket.mjs';
 import './legacy_ws7.mjs';
