+++ benchmark/servers/ports.mjs
@@ -3,5 +3,5 @@ export const ports = {
   ws7: 6543,
   uWebSockets: 6541,
   legacy_ws7: 6542,
-  'fastify-websocket_ws8': 6544,
+  '@fastify/websocket': 6544,
 };
