+++ docs/README.md
@@ -9,6 +9,7 @@ graphql-ws
 - [client](modules/client.md)
 - [common](modules/common.md)
 - [server](modules/server.md)
+- [use/@fastify/websocket](modules/use__fastify_websocket.md)
 - [use/fastify-websocket](modules/use_fastify_websocket.md)
 - [use/uWebSockets](modules/use_uWebSockets.md)
 - [use/ws](modules/use_ws.md)
